backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
psycopg[binary]==3.2.9
PyJWT==2.10.1
python-dotenv==1.0.1
supabase==2.15.0
Werkzeug==3.1.3
